+++ Src/Nager.Date.UnitTest/Country/PapuaNewGuineaTest.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Linq;
 
@@ -10,7 +10,7 @@ namespace Nager.Date.UnitTest.Country
         [TestMethod]
         public void CheckQueensBirthdayFor2018()
         {
-            var publicHolidays = DateSystem.GetPublicHoliday(2018, CountryCode.PG).ToArray();
+            var publicHolidays = DateSystem.GetPublicHolidays(2018, CountryCode.PG).ToArray();
 
             var publicHoliday = publicHolidays.FirstOrDefault(holiday => holiday.Name.Equals("Queen's Birthday"));
             Assert.AreEqual(new DateTime(2018, 6, 11), publicHoliday.Date);
@@ -19,7 +19,7 @@ namespace Nager.Date.UnitTest.Country
         [TestMethod]
         public void CheckQueensBirthdayFor2021()
         {
-            var publicHolidays = DateSystem.GetPublicHoliday(2021, CountryCode.PG).ToArray();
+            var publicHolidays = DateSystem.GetPublicHolidays(2021, CountryCode.PG).ToArray();
 
             var publicHoliday = publicHolidays.FirstOrDefault(holiday => holiday.Name.Equals("Queen's Birthday"));
             Assert.AreEqual(new DateTime(2021, 6, 14), publicHoliday.Date);