+++ Src/Nager.Date.UnitTest/Country/SwitzerlandTest.cs
@@ -1,4 +1,4 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Nager.Date.Extensions;
 using System;
 using System.Linq;
@@ -14,7 +14,7 @@ namespace Nager.Date.UnitTest.Country
         {
             for (var year = DateTime.Now.Year; year < 3000; year++)
             {
-                var publicHolidays = DateSystem.GetPublicHoliday(year, CountryCode.CH);
+                var publicHolidays = DateSystem.GetPublicHolidays(year, CountryCode.CH);
                 Assert.AreEqual(16, publicHolidays.Count());
             }
         }